move MDXProvider to blog-template.js
